Fix offset reference and move new tests to a separate, optional suite
