+++ ledger/ledger-api-test-tool/src/main/scala/com/daml/ledger/api/testtool/suites/CommandSubmissionCompletionIT.scala
@@ -3,24 +3,16 @@
 
 package com.daml.ledger.api.testtool.suites
 
-import com.daml.ledger.api.SubmissionIdGenerator
-import com.daml.ledger.api.testing.utils.MockMessages.offset
 import com.daml.ledger.api.testtool.infrastructure.Allocation._
 import com.daml.ledger.api.testtool.infrastructure.Assertions._
 import com.daml.ledger.api.testtool.infrastructure.LedgerTestSuite
-import com.daml.ledger.api.v1.command_submission_service.SubmitRequest
-import com.daml.ledger.api.v1.commands.Commands.DeduplicationPeriod
-import com.daml.ledger.api.v1.completion.Completion
 import com.daml.ledger.test.model.Test.Dummy
 import com.daml.ledger.test.model.Test.Dummy._
-import com.daml.lf.data.Ref
 import com.daml.platform.testing.{TimeoutException, WithTimeout}
-import com.google.protobuf.duration.Duration
 import io.grpc.Status
 
 import java.util.regex.Pattern
 import scala.concurrent.duration.DurationInt
-import scala.util.Try
 
 final class CommandSubmissionCompletionIT extends LedgerTestSuite {
 
@@ -170,155 +162,4 @@ final class CommandSubmissionCompletionIT extends LedgerTestSuite {
       // Nothing to do, if the two completions are found the test is passed
     }
   })
-
-  test(
-    "CSCIncludeApplicationId",
-    "The application ID is present in completions",
-    allocate(SingleParty),
-  )(implicit ec => { case Participants(Participant(ledger, party)) =>
-    val request = ledger.submitRequest(party, Dummy(party).create.command)
-    for {
-      _ <- ledger.submit(request)
-      completions <- ledger.firstCompletions(party)
-    } yield {
-      val actualCompletionApplicationId = singleCompletion(completions).applicationId
-      assert(
-        Ref.ApplicationId.fromString(actualCompletionApplicationId).contains(ledger.applicationId),
-        "Wrong application ID in completion, " +
-          s"expected: ${ledger.applicationId}, actual: $actualCompletionApplicationId",
-      )
-    }
-  })
-
-  test(
-    "CSCIncludeRequestedSubmissionId",
-    "The requested submission ID is present in the request's completion",
-    allocate(SingleParty),
-  )(implicit ec => { case Participants(Participant(ledger, party)) =>
-    val request = ledger.submitRequest(party, Dummy(party).create.command)
-    val requestedSubmissionId =
-      Some(Ref.SubmissionId.assertFromString(SubmissionIdGenerator.Random.generate()))
-    for {
-      _ <- ledger.submit(update(request, optSubmissionId = requestedSubmissionId))
-      completions <- ledger.firstCompletions(party)
-    } yield {
-      val actualCompletionSubmissionId = Option(singleCompletion(completions).submissionId)
-      assert(
-        actualCompletionSubmissionId == requestedSubmissionId,
-        "Wrong submission ID in completion, " +
-          s"expected: $requestedSubmissionId, actual: $actualCompletionSubmissionId",
-      )
-    }
-  })
-
-  test(
-    "CSCIncludeASubmissionIdWhenNotRequested",
-    "A completion includes a submission ID when one is missing in the request",
-    allocate(SingleParty),
-  )(implicit ec => { case Participants(Participant(ledger, party)) =>
-    val request = ledger.submitRequest(party, Dummy(party).create.command)
-    for {
-      _ <- ledger.submit(request)
-      completions <- ledger.firstCompletions(party)
-    } yield {
-      val actualCompletionSubmissionId = singleCompletion(completions).submissionId
-      assert(
-        Ref.SubmissionId.fromString(actualCompletionSubmissionId).isRight,
-        "Missing or invalid submission ID in completion",
-      )
-    }
-  })
-
-  test(
-    "CSCIncludeRequestedDeduplicationOffset",
-    "The requested deduplication offset is present in the request's completion",
-    allocate(SingleParty),
-  )(implicit ec => { case Participants(Participant(ledger, party)) =>
-    val request = ledger.submitRequest(party, Dummy(party).create.command)
-    val submittedDeduplication =
-      DeduplicationPeriod.DeduplicationOffset(offset(0).toHexString)
-    for {
-      _ <- ledger.submit(update(request, optDeduplicationPeriod = Some(submittedDeduplication)))
-      completions <- ledger.firstCompletions(party)
-    } yield {
-      val actualCompletionDeduplication = singleCompletion(completions).deduplicationPeriod
-      assert(
-        actualCompletionDeduplication.deduplicationOffset == submittedDeduplication.deduplicationOffset,
-        "Wrong duplication offset in completion, " +
-          s"expected: ${submittedDeduplication.deduplicationOffset}, " +
-          s"actual: ${actualCompletionDeduplication.deduplicationOffset}",
-      )
-    }
-  })
-
-  test(
-    "CSCIncludeRequestedDeduplicationTime",
-    "The requested deduplication time is present in the request's completion",
-    allocate(SingleParty),
-  )(implicit ec => { case Participants(Participant(ledger, party)) =>
-    val request = ledger.submitRequest(party, Dummy(party).create.command)
-    val submittedDeduplication =
-      DeduplicationPeriod.DeduplicationTime(Duration(seconds = 100, nanos = 10))
-    for {
-      _ <- ledger.submit(update(request, optDeduplicationPeriod = Some(submittedDeduplication)))
-      completions <- ledger.firstCompletions(party)
-    } yield {
-      val actualCompletionDeduplication = singleCompletion(completions).deduplicationPeriod
-      assert(
-        actualCompletionDeduplication.deduplicationTime == submittedDeduplication.deduplicationTime,
-        "Wrong duplication time in completion, " +
-          s"expected: ${submittedDeduplication.deduplicationTime}, " +
-          s"actual: ${actualCompletionDeduplication.deduplicationTime}",
-      )
-    }
-  })
-
-  test(
-    "CSCIncludeNoDeduplicationWhenNotRequested",
-    "A completion doesn't include a deduplication when one is missing in the request",
-    allocate(SingleParty),
-  )(implicit ec => { case Participants(Participant(ledger, party)) =>
-    val request = ledger.submitRequest(party, Dummy(party).create.command)
-    for {
-      _ <- ledger.submit(request)
-      completions <- ledger.firstCompletions(party)
-    } yield {
-      val actualCompletionDeduplication = singleCompletion(completions).deduplicationPeriod
-      assert(
-        actualCompletionDeduplication.isEmpty,
-        s"The deduplication $actualCompletionDeduplication " +
-          "is present in the completion even though it was not requested",
-      )
-    }
-  })
-
-  private def update(
-      submitRequest: SubmitRequest,
-      optSubmissionId: Option[Ref.SubmissionId] = None,
-      optDeduplicationPeriod: Option[DeduplicationPeriod] = None,
-  ): SubmitRequest = {
-    val optRequestUpdatedWithSubmission =
-      optSubmissionId.map { submissionId =>
-        submitRequest.copy(commands =
-          submitRequest.commands.map(_.copy(submissionId = submissionId))
-        )
-      }
-
-    val optRequestUpdatedWithDeduplicationPeriod =
-      optDeduplicationPeriod.map { deduplicationPeriod =>
-        optRequestUpdatedWithSubmission
-          .getOrElse(submitRequest)
-          .copy(commands =
-            submitRequest.commands.map(_.copy(deduplicationPeriod = deduplicationPeriod))
-          )
-      }
-
-    optRequestUpdatedWithDeduplicationPeriod.getOrElse(submitRequest)
-  }
-
-  private def singleCompletion(completions: Seq[Completion]): Completion =
-    assertSingleton(
-      "Expected exactly one completion",
-      completions,
-    )
 }
