+++ ledger/ledger-api-test-tool/src/main/scala/com/daml/ledger/api/testtool/tests/Tests.scala
@@ -61,6 +61,7 @@ object Tests {
   val optional: Vector[LedgerTestSuite] =
     Vector(
       new CommandDeduplicationOffsetIT,
+      new CommandSubmissionCompletionDeduplicationIT,
       new ContractIdIT,
       new MultiPartySubmissionIT,
       new ParticipantPruningIT,
